Pull Request (feature-ViewingAndSelectingObjects <- feature-RatingAndUser_management) : To update Frontend and Backend of User

+++ TrocAppBackend/src/main/java/helha/trocappbackend/repositories/RoleRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> , PagingAndSortingRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> , PagingAndSortingRepository<Role, Integer> {
     Optional<Role> findByName(String name);
     Optional<Role> findById(int id);
 }
